# docs/swagger.yaml
basePath: /
definitions:
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Reusable Go REST API backend template.
  title: WarSpark API
  version: dev
paths:
...
  /api/v1/clans/{tag}/war-log:
    get:
      parameters:
      - description: 部落标签
        in: path
        name: tag
        required: true
        type: string
      - description: 分页数量
        in: query
        name: limit
        type: integer
      - description: 游标后
        in: query
        name: after
        type: string
      - description: 游标前
        in: query
        name: before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落战争日志
      tags:
      - war
...
  /api/v1/war/current:
    get:
      parameters:
      - description: Clan tag
        in: query
        name: clan_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get current clan war
      tags:
      - war
  /api/v1/war/cwl:
    get:
      parameters:
      - description: Clan tag
        in: query
        name: clan_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get CWL group
      tags:
      - war
  /api/v1/war/cwl/wars/{war_tag}:
    get:
      parameters:
      - description: 战争标签
        in: path
        name: war_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取 CWL 联赛战争
      tags:
      - war
  /api/v1/war/snapshots/{war_snapshot_id}/members:
    get:
      parameters:
      - description: War snapshot ID
        in: path
        name: war_snapshot_id
        required: true
        type: string
      - description: clan or opponent
        in: query
        name: side
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List war snapshot members
      tags:
      - war
...
schemes:
- http
swagger: "2.0"